Multiple Joysticks support implemented
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@811 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/dialog/CDialog.h
@@ -40,7 +40,7 @@ private:
 	Uint16 m_w;
 	Uint16 m_h;
 
-	int m_selected_ID;
+	Uint8 m_selected_ID;
 	int m_dlg_size;
 	Uint8 m_switch;
 	Uint8 m_scroll;
